# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HaCkAtHoN_2025 API",
        "description": "API documentation for the HaCkAtHoN_2025",
        "contact": {
            "email": "support@yourapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/messages/": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Add a new message to a chat",
                "operationId": "e70221051b86213b7e9b9f415df6ea3f",
                "requestBody": {
                    "description": "Message data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "message",
                                    "is_ai",
                                    "chat_id",
                                    "type_id"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "What is your name?"
                                    },
                                    "is_ai": {
                                        "type": "inteager",
                                        "example": 0
                                    },
                                    "chat_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "successful": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Message created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "Message": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "message": {
                                                            "type": "string",
                                                            "example": "Hello Ouama"
                                                        },
                                                        "is_ai": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "type": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "type": {
                                                                    "type": "string",
                                                                    "example": "career"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "1 second ago"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid chat ID or type ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "successful": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The selected chat id is invalid."
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to create message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "successful": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized to create this Chat Message . This Chat Message isn't for you"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 403
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "successful": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "in": "header",
                "name": "Authorization",
                "description": "Enter your Bearer token in the format `YOUR_TOKEN`. For example: `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`"
            }
        }
    }
}